Face-Recognition/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==4.1.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy<2
tensorflow==2.15.0
mtcnn==0.1.1
scikit-learn==1.3.2
pillow==10.1.0
keras-facenet==0.3.2
imageio==2.31.5
scipy==1.11.4
